WebApplication/server/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
motor==3.3.2
python-jose[cryptography]==3.3.0
bcrypt==4.1.2
pandas==2.1.4
numpy==1.26.3
tensorflow==2.15.0
python-multipart==0.0.6
requests==2.31.0
scikit-learn==1.4.0
google-auth==2.27.0
google-auth-httplib2==0.2.0
pydantic-settings==2.1.0
httpx==0.26.0
python-dotenv==1.0.0
email-validator==2.1.0
